Add first version of simple note
Add first version of simple note using Python3 and it has a plus that it
only uses standard libraries.

It stores the list in persistent memory by saving the current state of
notes' storage in file `my-notes-backup-123` located in HOME. In theory
the program should run correctly in Linux and Windows, but it was only tested
in Linux.